Make container optional on linux hosts
