+++ pytest_vnet/__init__.py
@@ -1,3 +1,8 @@
 #!/usr/bin/env python3
 
-from .plugin import run_in_netvm, as_script, as_host
+from .vm import run_in_netvm
+from .plugin import (
+	as_script,
+	as_host,
+	vnet
+)
