+++ requirements.txt
@@ -1,0 +1,3 @@
+pytest
+docker
+git+git://github.com/mininet/mininet.git
